% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptBed.R
\name{AdaptBed}
\alias{AdaptBed}
\title{Adapt BED positions to RADseq simulation}
\usage{
AdaptBed(bed.file, out.file, int.size)
}
\arguments{
\item{bed.file}{character defining BED file}

\item{out.file}{character defining output BED file name}

\item{int.size}{interger defining threshold of interval size to be kept in dataset}
}
\description{
Adapt BED positions to RADseq simulation
}
