% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_founderfiles.R
\name{create_founderfiles}
\alias{create_founderfiles}
\title{Create PedigreeSim founder file}
\usage{
create_founderfiles(vcfR.object, parents_id, chrom)
}
\arguments{
\item{vcfR.object}{object of class vcfR}

\item{parents_id}{vector of character defining parents ID}

\item{chrom}{character defining the chromosome to be evaluated}
}
\description{
Create PedigreeSim founder file
}
