% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqToProfile.R
\name{seqToProfile}
\alias{seqToProfile}
\title{Create SimusCop profile}
\usage{
seqToProfile(
  bam.file = NULL,
  bed.file = NULL,
  vcf.file = NULL,
  reference = NULL,
  out.profile = NULL
)
}
\arguments{
\item{bam.file}{charater defining the BAM file}

\item{bed.file}{charater defining the BED file}

\item{vcf.file}{charater defining the VCF file}

\item{reference}{charater defining the reference genome FASTA file}

\item{out.profile}{charater defining the profile output name}
}
\description{
Create SimusCop profile
}
