% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuReads.R
\name{simuReads}
\alias{simuReads}
\title{Create SimusCop profile}
\usage{
simuReads(
  ref = NULL,
  profile = NULL,
  variation = NULL,
  target = NULL,
  name = NULL,
  output = NULL,
  layout = NULL,
  threads = NULL,
  verbose = NULL,
  coverage = NULL,
  insertSize = NULL
)
}
\arguments{
\item{ref}{fasta reference file from which reads will be sampled}

\item{profile}{a model file generated by "seqToProfile" utility}

\item{variation}{files defining the variations (indel, SNV and CNV) to simulate}

\item{target}{file defining target regions for sequencing. If the argument is not specified, sequencing data of all chromosomes defined in the reference will be produced.}

\item{name}{popolation names (comma-separated).}

\item{output}{output directory}

\item{layout}{sequence layout, should be "SE" for single end or "PE" for paired-end (default: "SE")}

\item{threads}{the number of threads to use (default: 1)}

\item{verbose}{print the intermediate information, should be 0 or 1 (default: 1)}

\item{coverage}{sequencing coverage}

\item{insertSize}{insert size for paired end sequencing (default: 350, only effective when the "layout" is set to "PE")}

\item{snp}{file defining the SNPs to simulate}
}
\description{
Create SimusCop profile
}
