% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2variants.R
\name{vcf2variants}
\alias{vcf2variants}
\title{Converts variants in VCF to simuscop input file.}
\usage{
vcf2variants(vcfR.object = NULL, sample = NULL, chrom = NULL)
}
\arguments{
\item{vcfR.object}{object of class vcfR}

\item{sample}{character defining sample ID to be evaluated}

\item{chrom}{character defining the chromosome to be evaluated}
}
\value{
list with data.frames for SNVs, indels and insertions
}
\description{
Exclusively for biallelic SNVs, indels and insertions. Only one sample at the time and one chromosome.
}
